requirements.txt
mcp
pandas
sqlalchemy
psycopg2-binary
python-dotenv
